requirements.txt
pillow==9.1.1
requests==2.28.0
google-api-python-client==2.95.0
python-dotenv==1.0.0
regex==2022.10.31